Add MSZIP decoder
Add a decoder for the MSZIP compression format.

There is some code-as-documentation about the protocol in my
reimplementation (<https://github.com/frereit/pymszip>), which I used
as a reference for this decoder.

This decoder does not actually perform any decompression, it only
highlights the relevant chunks and their headers. The individual chunks
are zlib compressed.